fix: stderr log lines getting squished together
Styles for span.log didn't apply when logs from stderr were wrapped
inside a <div> with the color. This change would make it so that *every*
log line would be wrappee with `span.log-line`, and the base styles are
applied for this. If the logs are from stderr, an extra class `.stderr`
would be added.

Signed-off-by: shrihanDev <shrihankp198725@gmail.com>